requirements.txt
google-cloud-bigquery
google-cloud-aiplatform
python-dotenv
pandas
sentence-transformers
fastapi
uvicorn
rapidfuzz
slack-sdk
aiohttp
db-dtypes
cachetools
sqlparse
python-dateutil
slack-bolt
pytest
python-multipart
